+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 
 - **[FIX]** Fix layout issue on `TopTripCardList`
+- **[NEW]** Add `rightTitleStrikeThrough` prop to `Item` and `ItemData`
+- [...]
 
 # v16.1.1 (12/12/2019)